+++ Entity/ProcessGroup.php
@@ -1,0 +1,111 @@
+<?php
+
+/*
+ * This file is part of aakb/kontrolgruppen-core-bundle.
+ *
+ * (c) 2019 ITK Development
+ *
+ * This source file is subject to the MIT license.
+ */
+
+namespace Kontrolgruppen\CoreBundle\Entity;
+
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
+use Doctrine\ORM\Mapping as ORM;
+
+/**
+ * @ORM\Entity(repositoryClass="Kontrolgruppen\CoreBundle\Repository\ProcessGroupRepository")
+ */
+class ProcessGroup
+{
+    /**
+     * @ORM\Id()
+     * @ORM\GeneratedValue()
+     * @ORM\Column(type="integer")
+     */
+    private $id;
+
+    /**
+     * @ORM\ManyToOne(targetEntity="Kontrolgruppen\CoreBundle\Entity\Process")
+     * @ORM\JoinColumn(nullable=false)
+     */
+    private $primaryProcess;
+
+    /**
+     * @ORM\ManyToMany(targetEntity="Kontrolgruppen\CoreBundle\Entity\Process", inversedBy="processGroups")
+     */
+    private $processes;
+
+    /**
+     * ProcessGroup constructor.
+     */
+    public function __construct()
+    {
+        $this->processes = new ArrayCollection();
+    }
+
+    /**
+     * @return int|null
+     */
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+
+    /**
+     * @return Process|null
+     */
+    public function getPrimaryProcess(): ?Process
+    {
+        return $this->primaryProcess;
+    }
+
+    /**
+     * @param Process|null $primaryProcess
+     *
+     * @return $this
+     */
+    public function setPrimaryProcess(?Process $primaryProcess): self
+    {
+        $this->primaryProcess = $primaryProcess;
+
+        return $this;
+    }
+
+    /**
+     * @return Collection|Process[]
+     */
+    public function getProcesses(): Collection
+    {
+        return $this->processes;
+    }
+
+    /**
+     * @param Process $process
+     *
+     * @return $this
+     */
+    public function addProcess(Process $process): self
+    {
+        if (!$this->processes->contains($process)) {
+            $this->processes[] = $process;
+        }
+
+        return $this;
+    }
+
+    /**
+     * @param Process $process
+     *
+     * @return $this
+     */
+    public function removeProcess(Process $process): self
+    {
+        if ($this->processes->contains($process)) {
+            $this->processes->removeElement($process);
+        }
+
+        return $this;
+    }
+}